Make the Java header generation stateless and thread-safe (#47)
The Java `ModulrApiAuth` implementation is currently persisting a `date`
and `lastUsedNonce` in fields, but without any synchronisation or
locking across the objects that update or use those fields. This could
result in signature being generated using a date, but that date then
being replaced by another thread, and the new date being used in the
`Date` header, so not matching the timestamp used in the signature.

To overcome this, `ModulrApiAuth` has been altered to be immutable, to
use nonces passed into the method, and to store dates locally rather
than in fields. Additionally, the implementation has been switched to
use Java time APIs to allow the use of thread-safe formatters, time-zone
aware handling of dates, and a Clock for supplying current datetime
instances.

Includes an upgrade to the latest major version of all dependencies.