Fix date formatting in Scala auth headers (#48)
The date format was using a UK locale so formatting September as `Sept`
rather than the expected `Sep` and therefore causing requests to be
rejected during signature checking. The code has been re-written to
follow a similar structure to the Java example, so using Clock, Instant,
and DateTimeFormat rather than Date and DateFormat, and specifies the
Locale as English as part of this to ensure month names are retained as
3 characters. Includes an upgrade to Scala 3 since this has been a
mainstream version for a few years.